+++ src/left_collapse.cpp
@@ -1,0 +1,23 @@
+#include <iostream>
+#include <vector>
+#include <algorithm>
+#include <unordered_map>
+#include <random>
+
+std::vector<char> left_collapse(char ch) {
+    std::unordered_map<char, std::vector<char>> possible_connections = {
+        {'d', {'l', 'd', 'u'}},
+        {'u', {'l', 'd', 'u'}},
+        {'l', {'b', 'r'}},
+        {'r', {'l', 'd', 'u'}},
+        {'b', {'b', 'r'}}
+    };
+
+    std::vector<char> options;
+
+    auto it = possible_connections.find(ch);
+
+    options = it->second;
+
+    return options;
+}
